+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Show number of days for available transfers in status widget
+
 ## [0.5.1] - 2021-03-07
 
 ### Added